# webapi/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "dixcord-bot webapi",
    "version": "3.0.0",
    "description": "Internal REST API backing the dixcord-bot Discord bot, webview, and webadmin services. Every route requires a service-account JWT (admin, bot, or webview) except login and health."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Service-account JWT obtained from POST /api/auth/login."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
      "AuthErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "description": "Shape returned by the authenticate/requireAdmin middleware, which omits `ok`.",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthErrorResponse"
            }
          }
        }
      },
      "ForbiddenRole": {
        "description": "Authenticated, but the account role is not permitted for this route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthErrorResponse"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Request body or query parameters failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/system/status": {
      "get": {
        "operationId": "getSystemStatus",
        "tags": [
          "System"
        ],
        "summary": "Get system and bot health status",
        "parameters": [
          {
            "name": "app",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "discord"
            },
            "description": "Together with guildId, scopes `status.bot` to one server. Omit for the most-recently-seen server across all of them."
          },
          {
            "name": "guildId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System status.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "status": {
                      "type": "object",
                      "properties": {
                        "webapi": {
                          "type": "string",
                          "example": "ok"
                        },
                        "db": {
                          "type": "string",
                          "example": "ok"
                        },
                        "dbType": {
                          "type": "string",
                          "enum": [
                            "mysql",
                            "sqlite"
                          ]
                        },
                        "cacheVersion": {
                          "type": "string"
                        },
                        "webapiUptimeSeconds": {
                          "type": "integer",
                          "description": "Seconds since this webapi process started."
                        },
                        "webapiMemoryRssBytes": {
                          "type": "integer",
                          "description": "Resident set size of this webapi process, in bytes."
                        },
                        "bot": {
                          "type": "object",
                          "nullable": true,
                          "description": "One server's status — see the `app`/`guildId` query params for how it's selected.",
                          "properties": {
                            "app": {
                              "type": "string"
                            },
                            "guildId": {
                              "type": "string"
                            },
                            "version": {
                              "type": "string"
                            },
                            "lastSeenAt": {
                              "type": "string"
                            },
                            "online": {
                              "type": "boolean"
                            },
                            "readyAt": {
                              "type": "string",
                              "nullable": true,
                              "description": "When the bot's current gateway session became ready; stable across heartbeats until the bot restarts."
                            },
                            "uptimeSeconds": {
                              "type": "integer",
                              "nullable": true,
                              "description": "Seconds since readyAt."
                            },
                            "memberCount": {
                              "type": "integer",
                              "nullable": true,
                              "description": "Guild member count as of the last heartbeat."
                            },
                            "channelCount": {
                              "type": "integer",
                              "nullable": true,
                              "description": "Cached guild channel count as of the last heartbeat."
                            },
                            "wsPingMs": {
                              "type": "integer",
                              "nullable": true,
                              "description": "Discord gateway heartbeat latency in ms as of the last heartbeat."
                            }
                          }
                        },
                        "bots": {
                          "type": "array",
                          "description": "Every known server's status, same shape as `bot`.",
                          "items": {
                            "type": "object",
                            "properties": {
                              "app": {
                                "type": "string"
                              },
                              "guildId": {
                                "type": "string"
                              },
                              "version": {
                                "type": "string"
                              },
                              "lastSeenAt": {
                                "type": "string"
                              },
                              "online": {
                                "type": "boolean"
                              },
                              "readyAt": {
                                "type": "string",
                                "nullable": true
                              },
                              "uptimeSeconds": {
                                "type": "integer",
                                "nullable": true
                              },
                              "memberCount": {
                                "type": "integer",
                                "nullable": true
                              },
                              "channelCount": {
                                "type": "integer",
                                "nullable": true
                              },
                              "wsPingMs": {
                                "type": "integer",
                                "nullable": true
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/system/invalidate-cache": {
      "post": {
        "operationId": "invalidateCache",
        "tags": [
          "System"
        ],
        "summary": "Increment cache version",
        "description": "Requires the admin role. Records an audit log entry.",
        "responses": {
          "200": {
            "description": "New cache version.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "cacheVersion": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenRole"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/system/heartbeat": {
      "post": {
        "operationId": "postHeartbeat",
        "tags": [
          "System"
        ],
        "summary": "Record a bot heartbeat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "app",
                  "guildId",
                  "version"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "enum": [
                      "discord"
                    ]
                  },
                  "guildId": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  },
                  "readyAt": {
                    "type": "string",
                    "nullable": true,
                    "description": "When the bot's current gateway session became ready. Send the same value on every heartbeat within one process lifetime; only changes on restart."
                  },
                  "memberCount": {
                    "type": "integer",
                    "nullable": true,
                    "description": "Guild member count."
                  },
                  "channelCount": {
                    "type": "integer",
                    "nullable": true,
                    "description": "Cached guild channel count."
                  },
                  "wsPingMs": {
                    "type": "integer",
                    "nullable": true,
                    "description": "Discord gateway heartbeat latency in ms."
                  },
                  "metrics": {
                    "type": "object",
                    "nullable": true,
                    "description": "Bot-internal cumulative counter snapshot (opaque; surfaced via GET /metrics)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Heartbeat recorded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}