# webapi/routes/system.js
import express from "express";
import { authenticate, requireAdmin } from "../middleware/auth.js";
import {
  getSystemStatus,
  getCacheVersion,
  incrementCacheVersion,
  recordBotHeartbeat,
} from "../services/systemStatus.js";
import { recordAudit } from "../services/auditLog.js";
import {
  CHAT_APP_PARAM_ERROR,
  resolveChatAppFromRequest,
} from "../utils/chatAppHttp.js";

const router = express.Router();

/**
 * GET /api/system/status
 * System and bot health status.
 * Query: { app?, guildId? } — when both given, `status.bot` is that server's status;
 * otherwise it's the most-recently-seen server across all of them (unchanged behavior).
 * Auth: required (admin, bot, or webview).
 * @openapi
 * /api/system/status:
 *   get:
 *     operationId: getSystemStatus
 *     tags: [System]
 *     summary: Get system and bot health status
 *     parameters:
 *       - name: app
 *         in: query
 *         required: false
 *         schema: { type: string, example: discord }
 *         description: Together with guildId, scopes `status.bot` to one server. Omit for the most-recently-seen server across all of them.
 *       - name: guildId
 *         in: query
 *         required: false
 *         schema: { type: string }
 *     responses:
 *       '200':
 *         description: System status.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 ok: { type: boolean, enum: [true] }
 *                 status:
 *                   type: object
 *                   properties:
 *                     webapi: { type: string, example: ok }
 *                     db: { type: string, example: ok }
 *                     dbType: { type: string, enum: [mysql, sqlite] }
 *                     cacheVersion: { type: string }
 *                     webapiUptimeSeconds: { type: integer, description: "Seconds since this webapi process started." }
 *                     webapiMemoryRssBytes: { type: integer, description: "Resident set size of this webapi process, in bytes." }
 *                     bot:
 *                       type: object
 *                       nullable: true
 *                       description: "One server's status — see the `app`/`guildId` query params for how it's selected."
 *                       properties:
 *                         app: { type: string }
 *                         guildId: { type: string }
 *                         version: { type: string }
 *                         lastSeenAt: { type: string }
 *                         online: { type: boolean }
 *                         readyAt: { type: string, nullable: true, description: "When the bot's current gateway session became ready; stable across heartbeats until the bot restarts." }
 *                         uptimeSeconds: { type: integer, nullable: true, description: "Seconds since readyAt." }
 *                         memberCount: { type: integer, nullable: true, description: "Guild member count as of the last heartbeat." }
 *                         channelCount: { type: integer, nullable: true, description: "Cached guild channel count as of the last heartbeat." }
 *                         wsPingMs: { type: integer, nullable: true, description: "Discord gateway heartbeat latency in ms as of the last heartbeat." }
 *                     bots:
 *                       type: array
 *                       description: Every known server's status, same shape as `bot`.
 *                       items:
 *                         type: object
 *                         properties:
 *                           app: { type: string }
 *                           guildId: { type: string }
 *                           version: { type: string }
 *                           lastSeenAt: { type: string }
 *                           online: { type: boolean }
 *                           readyAt: { type: string, nullable: true }
 *                           uptimeSeconds: { type: integer, nullable: true }
 *                           memberCount: { type: integer, nullable: true }
 *                           channelCount: { type: integer, nullable: true }
 *                           wsPingMs: { type: integer, nullable: true }
 *       '401':
 *         $ref: '#/components/responses/Unauthorized'
 *       '500':
 *         $ref: '#/components/responses/ServerError'
 */
router.get("/status", authenticate, async (req, res) => {
  try {
    const status = await getSystemStatus({
      app: typeof req.query.app === "string" ? req.query.app : undefined,
      guildId: typeof req.query.guildId === "string" ? req.query.guildId : undefined,
    });
    res.json({ ok: true, status });
  } catch (err) {
    console.error("GET /api/system/status error:", err);
    res.status(500).json({ ok: false, error: "Failed to get system status" });
  }
});

/**
 * GET /api/system/cache-version
 * Current cache version for bot polling.
 * Auth: required (bot or admin).
 * @openapi
 * /api/system/cache-version:
 *   get:
 *     operationId: getCacheVersion
 *     tags: [System]
 *     summary: Get current cache version
 *     responses:
 *       '200':
 *         description: Current cache version.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 ok: { type: boolean, enum: [true] }
 *                 cacheVersion: { type: string }
 *       '401':
 *         $ref: '#/components/responses/Unauthorized'
 *       '500':
 *         $ref: '#/components/responses/ServerError'
 */
router.get("/cache-version", authenticate, async (req, res) => {
  try {
    const cacheVersion = await getCacheVersion();
    res.json({ ok: true, cacheVersion });
  } catch (err) {
    console.error("GET /api/system/cache-version error:", err);
    res.status(500).json({ ok: false, error: "Failed to get cache version" });
  }
});

/**
 * POST /api/system/invalidate-cache
 * Increment cache version so bots reload cached content.
 * Auth: admin required.
 * @openapi
 * /api/system/invalidate-cache:
 *   post:
 *     operationId: invalidateCache
 *     tags: [System]
 *     summary: Increment cache version
 *     description: Requires the admin role. Records an audit log entry.
 *     responses:
 *       '200':
 *         description: New cache version.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 ok: { type: boolean, enum: [true] }
 *                 cacheVersion: { type: string }
 *       '401':
 *         $ref: '#/components/responses/Unauthorized'
 *       '403':
 *         $ref: '#/components/responses/ForbiddenRole'
 *       '500':
 *         $ref: '#/components/responses/ServerError'
 */
router.post("/invalidate-cache", authenticate, requireAdmin, async (req, res) => {
  try {
    const cacheVersion = await incrementCacheVersion();
    await recordAudit(req.user.id, "invalidate", "system_cache", null, {
      cacheVersion,
    });
    res.json({ ok: true, cacheVersion });
  } catch (err) {
    console.error("POST /api/system/invalidate-cache error:", err);
    res.status(500).json({ ok: false, error: "Failed to invalidate cache" });
  }
});

/**
 * POST /api/system/heartbeat
 * Bot heartbeat (app, guild id, version, and optional session/health fields).
 * Body: { app, guildId, version, readyAt?, memberCount?, channelCount?, wsPingMs? }
 * Auth: required (bot or admin).
 * @openapi
 * /api/system/heartbeat:
 *   post:
 *     operationId: postHeartbeat
 *     tags: [System]
 *     summary: Record a bot heartbeat
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [app, guildId, version]
 *             properties:
 *               app: { type: string, enum: [discord] }
 *               guildId: { type: string }
 *               version: { type: string }
 *               readyAt: { type: string, nullable: true, description: "When the bot's current gateway session became ready. Send the same value on every heartbeat within one process lifetime; only changes on restart." }
 *               memberCount: { type: integer, nullable: true, description: "Guild member count." }
 *               channelCount: { type: integer, nullable: true, description: "Cached guild channel count." }
 *               wsPingMs: { type: integer, nullable: true, description: "Discord gateway heartbeat latency in ms." }
 *               metrics: { type: object, nullable: true, description: "Bot-internal cumulative counter snapshot (opaque; surfaced via GET /metrics)." }
 *     responses:
 *       '200':
 *         description: Heartbeat recorded.
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 ok: { type: boolean, enum: [true] }
 *       '400':
 *         $ref: '#/components/responses/BadRequest'
 *       '401':
 *         $ref: '#/components/responses/Unauthorized'
 *       '500':
 *         $ref: '#/components/responses/ServerError'
 */
router.post("/heartbeat", authenticate, async (req, res) => {
  try {
    const app = resolveChatAppFromRequest(req);
    if (!app) return res.status(400).json(CHAT_APP_PARAM_ERROR);
    const guildId = String(req.body?.guildId ?? "").trim();
    const version = String(req.body?.version ?? "").trim();
    if (!guildId || !version) {
      return res.status(400).json({
        ok: false,
        error: "guildId and version are required",
      });
    }
    await recordBotHeartbeat({
      app,
      guildId,
      version,
      readyAt: req.body?.readyAt,
      memberCount: req.body?.memberCount,
      channelCount: req.body?.channelCount,
      wsPingMs: req.body?.wsPingMs,
      metrics: req.body?.metrics,
    });
    res.json({ ok: true });
  } catch (err) {
    console.error("POST /api/system/heartbeat error:", err);
    res.status(500).json({ ok: false, error: "Failed to record heartbeat" });
  }
});

export default router;